test(shell_allowlist): split substitution tests into their own file
The lock additions from the previous commit pushed tests.rs to 1505
lines, over the repo's 1500-line LOC gate (scripts/loc-gate.sh), which
explicitly forbids growing the allowlist back — legacy files must be
split, not exempted.

Extract the 5 command-substitution tests (gh391_strict_mode_blocks_
substitution_in_args and the 4 substitution_with_* tests) into
substitution_tests.rs, following the same #[path]-based submodule
pattern already used for comment_strip_tests.rs. Behavior and test
names are unchanged.
